+++ includes/utils.h
@@ -6,7 +6,7 @@
 /*   By: totaisei <totaisei@student.42tokyo.jp>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/02/12 20:13:16 by nfukada           #+#    #+#             */
-/*   Updated: 2021/02/25 14:42:30 by totaisei         ###   ########.fr       */
+/*   Updated: 2021/02/25 15:34:51 by totaisei         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -22,6 +22,15 @@ typedef struct	s_env
 	char			*value;
 	struct s_env	*next;
 }				t_env;
+
+typedef enum	e_cmd_type
+{
+	ABSOLUTE,
+	RELATIVE,
+	COMMAND
+}				t_cmd_type;
+
+typedef struct stat t_stat;
 
 void			error_exit(void);
 void			print_unexpected_token_error(t_token *token);
